delete `parse` module, combine with `parser`
 * move `::parse_program` and `::parse_expression` to `parser` module
 * make `::Parser` class private to the `parser` module (move to .cpp file)
 * add a missing include directive in `token_array.hpp`
