+++ src/token_array.hpp
@@ -2,4 +2,6 @@
 
 #include "utils/chunked_array.hpp"
 
+struct Token;
+
 using TokenArray = ChunkedArray<Token>;
